Split canal ESIL analysis into include file
